Issue 799. Remove KerasWrapper in favor of raw KerasModel
