+++ deeppavlov/core/models/keras_model.py
@@ -12,13 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import inspect
 from abc import abstractmethod
-from copy import deepcopy
 from logging import getLogger
-from typing import Optional, List, Union
 
-import numpy as np
 import tensorflow as tf
 from keras import backend as K
 from overrides import overrides
@@ -94,86 +90,6 @@ class KerasModel(NNModel, metaclass=TfModelMeta):
             self.batches_seen = data["batches_seen"]
             self.train_examples_seen = data["train_examples_seen"]
         return
-
-
-class KerasWrapper(KerasModel):
-    """A wrapper over external Keras models. It is used, for example,
-    to wrap :class:`~deeppavlov.models.morpho_tagger.network.CharacterTagger`.
-    A subclass of :class:`~deeppavlov.core.models.keras_model.KerasModel`
-
-    Attributes:
-        cls: the class to be wrapped
-        save_path: the path where model is saved
-        load_path: the path from where model is loaded
-        mode: usage mode
-        **kwargs: a dictionary containing model parameters specified in the main part
-            of json config that corresponds to the model
-    """
-    def __init__(self, cls: type, save_path: Optional[str] = None,
-                 load_path: Optional[str] = None, mode: str = None,
-                 **kwargs) -> None:
-        # Calls parent constructor. Results in creation of save_folder if it doesn't exist
-        super().__init__(save_path=save_path, load_path=load_path, mode=mode)
-
-        # Dicts are mutable! To prevent changes in config dict outside this class
-        # we use deepcopy
-        opt = deepcopy(kwargs)
-
-        # Finds all input parameters of the network __init__ to pass them into network later
-        network_parameter_names = list(inspect.signature(cls.__init__).parameters)
-        # Fills all provided parameters from opt (opt is a dictionary formed from the model
-        # json config file, except the "name" field)
-        network_parameters = {par: opt[par] for par in network_parameter_names if par in opt}
-        self._net = cls(**network_parameters)
-
-        # Finds all parameters for network train to pass them into train method later
-        train_parameters_names = list(inspect.signature(self._net.train_on_batch).parameters)
-
-        # Fills all provided parameters from opt
-        train_parameters = {par: opt[par] for par in train_parameters_names if par in opt}
-        self.train_parameters = train_parameters
-        self.opt = opt
-
-        # Tries to load the model from model `load_path`, if it is available
-        self.load()
-
-    def load(self) -> None:
-        """Checks existence of the model file, loads the model if the file exists"""
-
-        # Checks presence of the model files
-        if self.load_path.exists():
-            path = str(self.load_path.resolve())
-            log.info('[loading model from {}]'.format(path))
-            self._net.load(path)
-
-    def save(self) -> None:
-        """Saves model to the save_path, provided in config. The directory is
-        already created by super().__init__, which is called in __init__ of this class"""
-        path = str(self.save_path.absolute())
-        log.info('[saving model to {}]'.format(path))
-        self._net.save(path)
-
-    def train_on_batch(self, *args) -> None:
-        """Trains the model on a single batch.
-
-        Args:
-            *args: the list of network inputs.
-            Last element of `args` is the batch of targets,
-            all previous elements are training data batches
-        """
-        *data, labels = args
-        self._net.train_on_batch(data, labels)
-
-    def __call__(self, *x_batch, **kwargs) -> Union[List, np.ndarray]:
-        """
-        Predicts answers on batch elements.
-
-        Args:
-            instance: a batch to predict answers on
-        """
-        with self.graph.as_default():
-            K.set_session(self.sess)
-            return self._net.predict_on_batch(x_batch, **kwargs)
 
 
 class LRScheduledKerasModel(LRScheduledModel, KerasModel):
